refactored regrid to put common files in separate module
