+++ tests/test_python/test_smoke_dust/conftest.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 import netCDF4 as nc
 import numpy as np
-
 import pytest
 
 from smoke_dust.core.context import SmokeDustContext
